refactor/함수, 상수 분리해서 리팩토링
