+++ src/components/common/header/Cart.tsx
@@ -1,17 +1,9 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import { FaShoppingCart } from "react-icons/fa";
-import { useRecoilValue } from "recoil";
-import { cartAtom } from "../../../atoms";
+import { getTotalQuantity } from "../../../shared/utils";
 
 function Cart() {
-  const cart = useRecoilValue(cartAtom);
-  const getTotalQuantity = () => {
-    return cart.reduce((acc, cur) => {
-      return acc + cur.quantity;
-    }, 0);
-  };
-
   return (
     <div className="hover:bg-gray-100 px-3 rounded">
       <Link className="px-5" to="/cart">
