+++ src/shared/constants.tsx
@@ -1,0 +1,3 @@
+const STAR_NUM = 5;
+
+export { STAR_NUM };
